+++ CESystem/ComponentRender.h
@@ -9,22 +9,22 @@ private:
 	std::string tileset;
 	sf::Vector2u tileSize;
 	sf::Vector2u tilePosition;
-	sf::Vector2i position;
-	sf::Vector2i oldPosition;
+	sf::Vector2f position;
+	sf::Vector2f oldPosition;
 	
 public:
-	ComponentRender(std::string tileset, sf::Vector2u tileSize, sf::Vector2u tilePosition, sf::Vector2i position);
+	ComponentRender(std::string tileset, sf::Vector2u tileSize, sf::Vector2u tilePosition, sf::Vector2f position);
 	ComponentRender(bool(&ckey)[components::SIZE]);
 	~ComponentRender();
 
 	std::string getTileset();
 	sf::Vector2u getTileSize();
 	sf::Vector2u getTilePosition();
-	sf::Vector2i getPosition();
-	int getPositionX();
-	int getPositionY();
+	sf::Vector2f getPosition();
+	float getPositionX();
+	float getPositionY();
 
-	void setPosition(sf::Vector2i position);
+	void setPosition(sf::Vector2f position);
 	void setTileSize(sf::Vector2u tileSize);
 	void setTileset(std::string tileset);
 	void setTilePosition(sf::Vector2u tilePosition);
